+++ lbrynet/reflector/server/server.py
@@ -4,7 +4,7 @@ from twisted.python import failure
 from twisted.internet import error, defer
 from twisted.internet.protocol import Protocol, ServerFactory
 from lbrynet.core.utils import is_valid_blobhash
-from lbrynet.core.Error import DownloadCanceledError, InvalidBlobHashError, NoSuchSDHash
+from lbrynet.core.Error import DownloadCanceledError, InvalidBlobHashError
 from lbrynet.core.StreamDescriptor import BlobStreamDescriptorReader
 from lbrynet.core.StreamDescriptor import save_sd_info
 from lbrynet.reflector.common import REFLECTOR_V1, REFLECTOR_V2
@@ -66,36 +66,6 @@ class ReflectorServer(Protocol):
             log.exception(err)
 
     @defer.inlineCallbacks
-    def check_head_blob_announce(self, stream_hash):
-        head_blob_hash = yield self.storage.get_stream_blob_by_position(stream_hash, 0)
-        if head_blob_hash in self.blob_manager.blobs:
-            head_blob = self.blob_manager.blobs[head_blob_hash]
-            if head_blob.get_is_verified():
-                should_announce = yield self.blob_manager.get_should_announce(head_blob_hash)
-                if should_announce == 0:
-                    yield self.blob_manager.set_should_announce(head_blob_hash, 1)
-                    log.info("Discovered previously completed head blob (%s), "
-                             "setting it to be announced", head_blob_hash[:8])
-        defer.returnValue(None)
-
-    @defer.inlineCallbacks
-    def check_sd_blob_announce(self, sd_hash):
-        if sd_hash in self.blob_manager.blobs:
-            sd_blob = self.blob_manager.blobs[sd_hash]
-            if sd_blob.get_is_verified():
-                should_announce = yield self.blob_manager.get_should_announce(sd_hash)
-                if should_announce == 0:
-                    yield self.blob_manager.set_should_announce(sd_hash, 1)
-                    log.info("Discovered previously completed sd blob (%s), "
-                             "setting it to be announced", sd_hash[:8])
-                    stream_hash = yield self.storage.get_stream_hash_for_sd_hash(sd_hash)
-                    if not stream_hash:
-                        log.info("Adding blobs to stream")
-                        sd_info = yield BlobStreamDescriptorReader(sd_blob).get_info()
-                        yield save_sd_info(self.blob_manager, sd_hash, sd_info)
-        defer.returnValue(None)
-
-    @defer.inlineCallbacks
     def _on_completed_blob(self, blob, response_key):
         yield self.blob_manager.blob_completed(blob, should_announce=False)
 
@@ -103,27 +73,15 @@ class ReflectorServer(Protocol):
             sd_info = yield BlobStreamDescriptorReader(blob).get_info()
             yield save_sd_info(self.blob_manager, blob.blob_hash, sd_info)
             yield self.blob_manager.set_should_announce(blob.blob_hash, True)
-
-            # if we already have the head blob, set it to be announced now that we know it's
-            # a head blob
-            d = self.check_head_blob_announce(sd_info['stream_hash'])
-
         else:
-            d = defer.succeed(None)
             stream_hash = yield self.storage.get_stream_of_blob(blob.blob_hash)
             if stream_hash is not None:
                 blob_num = yield self.storage.get_blob_num_by_hash(stream_hash,
                                                                    blob.blob_hash)
                 if blob_num == 0:
-                    sd_hash = yield self.storage.get_sd_blob_hash_for_stream(stream_hash)
                     yield self.blob_manager.set_should_announce(blob.blob_hash, True)
 
-                    # if we already have the sd blob, set it to be announced now that we know it's
-                    # a sd blob
-                    d.addCallback(lambda _: self.check_sd_blob_announce(sd_hash))
-
         yield self.close_blob()
-        yield d
         log.info("Received %s", blob)
         yield self.send_response({response_key: True})
 
@@ -238,7 +196,6 @@ class ReflectorServer(Protocol):
         if int(request_dict[VERSION]) not in [REFLECTOR_V1, REFLECTOR_V2]:
             raise ReflectorClientVersionError("Unknown version: %i" % int(request_dict[VERSION]))
 
-
         self.peer_version = int(request_dict[VERSION])
         log.debug('Handling handshake for client version %i', self.peer_version)
         self.received_handshake = True
@@ -292,17 +249,9 @@ class ReflectorServer(Protocol):
     @defer.inlineCallbacks
     def get_descriptor_response(self, sd_blob):
         if sd_blob.get_is_verified():
-            # if we already have the sd blob being offered, make sure we have it and the head blob
-            # marked as such for announcement now that we know it's an sd blob that we have.
-            yield self.check_sd_blob_announce(sd_blob.blob_hash)
-            try:
-                stream_hash = yield self.storage.get_stream_hash_for_sd_hash(
-                    sd_blob.blob_hash)
-            except NoSuchSDHash:
-                sd_info = yield BlobStreamDescriptorReader(sd_blob).get_info()
-                stream_hash = sd_info['stream_hash']
-                yield save_sd_info(self.blob_manager, sd_blob.blob_hash, sd_info)
-            yield self.check_head_blob_announce(stream_hash)
+            sd_info = yield BlobStreamDescriptorReader(sd_blob).get_info()
+            yield save_sd_info(self.blob_manager, sd_blob.blob_hash, sd_info)
+            yield self.storage.verify_will_announce_head_and_sd_blobs(sd_info['stream_hash'])
             response = yield self.request_needed_blobs({SEND_SD_BLOB: False}, sd_blob)
         else:
             self.incoming_blob = sd_blob
@@ -311,35 +260,11 @@ class ReflectorServer(Protocol):
             response = {SEND_SD_BLOB: True}
         defer.returnValue(response)
 
+    @defer.inlineCallbacks
     def request_needed_blobs(self, response, sd_blob):
-        def _add_needed_blobs_to_response(needed_blobs):
-            response.update({NEEDED_BLOBS: needed_blobs})
-            return response
-
-        d = self.determine_missing_blobs(sd_blob)
-        d.addCallback(_add_needed_blobs_to_response)
-        return d
-
-    def determine_missing_blobs(self, sd_blob):
-        reader = sd_blob.open_for_reading()
-        sd_blob_data = reader.read()
-        reader.close()
-        decoded_sd_blob = json.loads(sd_blob_data)
-        return self.get_unvalidated_blobs_in_stream(decoded_sd_blob)
-
-    def get_unvalidated_blobs_in_stream(self, sd_blob):
-        dl = defer.DeferredList(list(self._iter_unvalidated_blobs_in_stream(sd_blob)),
-                                consumeErrors=True)
-        dl.addCallback(lambda needed: [blob[1] for blob in needed if blob[1]])
-        return dl
-
-    def _iter_unvalidated_blobs_in_stream(self, sd_blob):
-        for blob in sd_blob['blobs']:
-            if 'blob_hash' in blob and 'length' in blob:
-                blob_hash, blob_len = blob['blob_hash'], blob['length']
-                d = self.blob_manager.get_blob(blob_hash, blob_len)
-                d.addCallback(lambda blob: blob_hash if not blob.get_is_verified() else None)
-                yield d
+        needed_blobs = yield self.storage.get_pending_blobs_for_stream(sd_blob.blob_hash)
+        response.update({NEEDED_BLOBS: needed_blobs})
+        defer.returnValue(response)
 
     def handle_blob_request(self, request_dict):
         """