+++ CHANGELOG.md
@@ -15,14 +15,16 @@ at anytime.
 ### Fixed
   * fixed the inconsistencies in API and CLI docstrings
   * `blob_announce` error when announcing a single blob
+  * `blob_list` error when looking up blobs by stream or sd hash
 
 ### Deprecated
   * `report_bug` jsonrpc command
   *
 
 ### Changed
-  *
-  *
+  * reflector server to periodically check and set `should_announce` for sd and head blobs instead of during each request
+  * reflector server to use `SQLiteStorage` to find needed blob hashes for a stream
+
 ### Added
   * scripts to autogenerate documentation
   *